# backend-go/openapi.yaml
openapi: 3.0.3
info:
  title: Cancer Detection Backend API
  version: 1.0.0
  description: |
    Gateway-facing backend API.
    Internal microservice communication may use gRPC, but external clients should use these HTTP endpoints.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/images:
    get:
      tags: [Images]
      summary: List images
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  images:
                    type: array
                    items: { type: object, additionalProperties: true }
                  nextCursor: { type: string, nullable: true }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/images/upload:
    post:
      tags: [Images]
      summary: Upload image and trigger AI detection
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                consentAccepted:
                  type: string
                  example: "true"
                consentVersion:
                  type: string
                  example: v1.0
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  image:
                    type: object
                    additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    BadGateway:
      description: Bad Gateway
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true
...